Create advanced-pull-to-scene
Macro Title: Advanced Pull to Scene
Author: TheJoester (https://github.com/thejoester)
Description: This macro allows a FoundryVTT GM to pull selected connected users to a chosen scene. It presents a dialog box listing all connected users with checkboxes, allowing the GM to select specific users. The dialog also includes a dropdown menu listing all available scenes, from which the GM can choose the target scene to pull users to. When the "Pull to Scene" button is clicked, the selected users are transported to the selected scene.

Usage: Run this macro as a GM in FoundryVTT.

Foundry Version: 12
Last updated 12-Nov-2024

Author: TheJoester (https://github.com/thejoester)
License: MIT License